chore: remove obsolete scripts and notebooks (#641)
* Delete add_users.py

Deleting, as agreed on with Edouard. This content is already present in documentation.

* Delete api_key.md

Deleting, as agreed on with Edouard. This content is already present in documentation.

* Delete create_project.ipynb

Deleting, as agreed on with Edouard. This content is already present in documentation.

* Delete efficient_data_augmentation.ipynb

Deleting, as agreed on with Edouard. This content is already present in documentation.

* Delete entrypoint.sh

Deleting, as agreed on with Edouard. This content is already present in documentation.

* Delete export_labels.ipynb

Deleting, as agreed on with Edouard. This content is already present in documentation.

* Delete import_assets.ipynb

Deleting, as agreed on with Edouard. This content is already present in documentation.

* Delete kili_to_pascal_voc.py

Exporting to Pascal VOC already part of standard Kili offering. Removing, as agreed on with Edouard.

* Delete delete_all_assets.py

Moved this content to documentation as a new recipe. Deleting from here, per conversation with Edouard

* Delete set_asset_to_be_labeled_by.py

This is now converted to a recipe in documentation:
https://docs.kili-technology.com/recipes/assigning-asset-to-user

Deleting from here, per conversation with Edouard.

* Delete update_project_instructions.py

This is now a recipe in documentation:
https://docs.kili-technology.com/recipes/setting-project-instructions

Removing from here, per conversation with Edouard

* Delete update_project_title.py

This is now a recipe here:
https://docs.kili-technology.com/recipes/updating-project-title

Removing from here, per conversation with Edouard

* Delete import_predictions.ipynb

This nb is 1) outdated and 2) repeats what's already described in documentation:
- https://docs.kili-technology.com/docs/importing-labels
- https://docs.kili-technology.com/docs/importing-labels
- https://docs.kili-technology.com/reference/export-classification
- https://docs.kili-technology.com/reference/export-object-entity-detection-and-relation
- https://docs.kili-technology.com/reference/export-transcription

Removing this nb from here, per conversation with Edouard